Iteration2 future 189

+++ src/main/java/ru/innopolis/stc16/innobazaar/dao/UserDAO.java
@@ -9,9 +9,9 @@ public interface UserDAO {
 
     public void saveUser(User user);
 
-    public User getUser(int id);
+    public User getUser(Long id);
 
-    public void deleteUser(int id);
+    public void deleteUser(Long id);
 
     public void updateUser(User user);
