+++ src/main/java/ru/innopolis/stc16/innobazaar/service/UserService.java
@@ -10,9 +10,9 @@ public interface UserService {
 
     public void saveUser(User user);
 
-    public User getUser(int id);
+    public User getUser(Long id);
 
-    public void deleteUser(int id);
+    public void deleteUser(Long id);
 
     public void updateUser(User user);
